Start splitting codegen to improve code clarity
